requirements.txt
aiohttp==3.7.3
async_timeout==3.0.1
colorama==0.4.4
discord.py==1.7.3
python-dotenv==0.20.0
requests==2.25.1
youtube_dl==2021.12.17
